fix: split data.ts to prevent server code leaking into client bundle
Moving getRankings and DB deps to get-rankings.ts keeps drizzle-orm
and pg (Node builtins: dns, fs, net, tls) out of the client bundle
when RankingsPage.tsx ("use client") imports format utilities from data.ts.
